♻️ Refactor address handling: consolidate address utilities and remove deprecated shortenAddress function
